Deploy: ver.0.0.0 (Test)

+++ src/components/detailModal/LikeBtn.tsx
@@ -5,6 +5,10 @@ import usePostLikeMutation from '../../hooks/reactQuery/like/usePostLikeMutation
 import useDeleteLikeMutation from '../../hooks/reactQuery/like/useDeleteLikeMutation';
 import useGetLikeQuery from '../../hooks/reactQuery/like/useGetLikeQuery';
 import { LikePostsType } from '../../model/interface';
+import AlertModal from '../Modal/AlertModal';
+import CommentError from '../share/alert/CommentError';
+import useAlertModalCtr from '../../hooks/useAlertModalCtr';
+import _ from 'lodash';
 
 interface Props {
   postId: number | undefined;
@@ -14,7 +18,9 @@ const LikeBtn: React.FC<Props> = ({ postId }) => {
   const { data, isLoading } = useGetLikeQuery();
   const [isLiked, setIsLiked] = useState(false);
   const [isHover, setIsHover] = useState(true);
+  const { isModalOpen, handleModalClose, handleModalOpen } = useAlertModalCtr();
   const [scope, animate] = useAnimate();
+  const userId = window.sessionStorage.getItem('userId');
 
   useEffect(() => {
     if (!isLoading) {
@@ -22,20 +28,36 @@ const LikeBtn: React.FC<Props> = ({ postId }) => {
     }
   }, [data, isLoading, postId]);
 
-  const { postMutate, isPostLoading } = usePostLikeMutation(postId);
+  const { postMutate, isPostLoading } = usePostLikeMutation(
+    postId,
+    handleModalOpen,
+    setIsLiked,
+  );
   const { deleteMutate, isDeleteLoading } = useDeleteLikeMutation(postId);
 
-  const handleLikeClick = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    e.stopPropagation();
-    animate([]);
-    if (isLiked) {
-      deleteMutate({ postId });
-      setIsLiked(false);
-    } else {
-      postMutate({ postId });
-      setIsLiked(true);
-    }
-  };
+  const handleLikeClick = _.throttle(
+    (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
+      e.stopPropagation();
+      animate([]);
+
+      if (isPostLoading || isDeleteLoading) {
+        return;
+      }
+
+      if (!userId) return handleModalOpen();
+
+      if (isLiked) {
+        deleteMutate({ postId });
+        setIsLiked(false);
+        return;
+      } else {
+        postMutate({ postId });
+        setIsLiked(true);
+        return;
+      }
+    },
+    400,
+  );
 
   return (
     <LikeBtnWrapper
@@ -74,6 +96,13 @@ const LikeBtn: React.FC<Props> = ({ postId }) => {
           )}
         </svg>
       </MotionDiv>
+      <AlertModal
+        isOpen={isModalOpen}
+        onClose={handleModalClose}
+        position="topRight"
+      >
+        <CommentError />
+      </AlertModal>
     </LikeBtnWrapper>
   );
 };
